Introduce `@EnableMessagingAnnotations` to activate messaging annotations post processors
this is another option for #11024
